Query the database asynchronously and schedule changes
"schedule changes" means that it uses Sponge.server().scheduler() to
queue the events to be done when the server is able to. This is in an
attempt to not lag the entire server when rollbacks and restores are done.
No idea if it'll work honestly

Squashed commit of the following: