Fix synchronization of record rollbacks and restores
Sort by id descending for rollbacks, and id ascending for restores.

Also iterates records on the Sponge scheduler (synchronously) so that we always
rollback in the correct order.

Fixes many issues related to rollback/restore/undo/redo, it seems.

Also adds a debug config option with messages, and a --sort flag
to transaction commands (only if debug mode is on.)